Add QDQ stripping to separate files
